+++ src/pages/sponsors/sponsor-reports/sponsor-asset-drilldown-page/index.js
@@ -41,8 +41,7 @@ import PrintIcon from "@mui/icons-material/Print";
 import ImageOutlinedIcon from "@mui/icons-material/ImageOutlined";
 import InsertDriveFileOutlinedIcon from "@mui/icons-material/InsertDriveFileOutlined";
 import DownloadIcon from "@mui/icons-material/Download";
-import { htmlToPlainText } from "../../../../utils/methods";
-import { isPositiveIntId } from "../../../../utils/reports-api";
+import { htmlToPlainText, isPositiveIntId } from "../../../../utils/methods";
 import ReportShell from "../../../../components/sponsors/reports/ReportShell";
 import usePrint from "../../../../hooks/usePrint";
 import TierBadge from "../../../../components/sponsors/reports/TierBadge";
